Player can move in 8 directions and doesn't need to be in the middle of a tile

+++ src/main/Map.java
@@ -7,8 +7,8 @@ public class Map {
 
 	private final TiledMap tiledMap;
 	
-	private int x = 0;
-	private int y = 0;
+	private float x = 0;
+	private float y = 0;
 	
 	public Map(String path) throws SlickException {
 		
@@ -17,30 +17,30 @@ public class Map {
 	}
 	
 	public void render() {
-		
-		tiledMap.render(x, y);
+				
+		tiledMap.render((int) x, (int) y);
 		
 	}
 
-	public int getX() {
+	public float getX() {
 		
 		return x;
 		
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		
 		this.x = x;
 		
 	}
 
-	public int getY() {
+	public float getY() {
 		
 		return y;
 		
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		
 		this.y = y;
